+++ test/next/src/pages/api/vercel/blob/pages/serverless.ts
@@ -1,5 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import * as vercelBlob from '@vercel/blob';
+import { validateUploadToken } from '@/app/vercel/blob/validate-upload-token';
 
 export const config = {
   runtime: 'nodejs',
@@ -13,6 +14,10 @@ export default async function handleBody(
 
   if (!request.body || !pathname) {
     return response.status(400).json({ message: 'No file to upload.' });
+  }
+
+  if (!validateUploadToken(request)) {
+    return response.status(401).json({ message: 'Not authorized' });
   }
 
   // Note: this will stream the file to Vercel's Blob Store